approaches/code-warriors2/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=2.1.0=pyhd8ed1ab_0
anyio=4.3.0=pyhd8ed1ab_0
argon2-cffi=23.1.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py312h98912ed_4
arrow=1.3.0=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
async-lru=2.0.4=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
babel=2.14.0=pyhd8ed1ab_0
beautifulsoup4=4.12.3=pyha770c72_0
blas=2.116=mkl
blas-devel=3.9.0=16_linux64_mkl
bleach=6.1.0=pyhd8ed1ab_0
brotli-python=1.1.0=py312h30efb56_1
bzip2=1.0.8=hd590300_5
c-ares=1.28.1=hd590300_0
ca-certificates=2024.8.30=hbcca054_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.16.0=py312hf06ca03_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.99=0
cuda-runtime=12.1.0=0
cycler=0.12.1=pypi_0
debugpy=1.8.1=py312h30efb56_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pyhd8ed1ab_0
fonttools=4.50.0=pypi_0
fqdn=1.5.1=pyhd8ed1ab_0
freetype=2.12.1=h267a509_2
gmp=6.3.0=h59595ed_1
gnutls=3.6.13=h85f3911_1
grpcio=1.62.2=py312hb06c811_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.4=pyhd8ed1ab_0
httpx=0.27.0=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=73.2=h59595ed_0
idna=3.6=pyhd8ed1ab_0
imageio=2.34.0=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
importlib_resources=6.3.2=pyhd8ed1ab_0
ipykernel=6.29.3=pyhd33586a_0
ipython=8.22.2=pyh707e725_0
ipywidgets=8.1.2=pyhd8ed1ab_0
isoduration=20.11.0=pyhd8ed1ab_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pyhd8ed1ab_0
joblib=1.3.2=pypi_0
json5=0.9.24=pyhd8ed1ab_0
jsonpointer=2.4=py312h7900ff3_3
jsonschema=4.21.1=pyhd8ed1ab_0
jsonschema-specifications=2023.12.1=pyhd8ed1ab_0
jsonschema-with-format-nongpl=4.21.1=pyhd8ed1ab_0
jupyter=1.0.0=pyhd8ed1ab_10
jupyter-lsp=2.2.4=pyhd8ed1ab_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_console=6.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h7900ff3_0
jupyter_events=0.10.0=pyhd8ed1ab_0
jupyter_server=2.13.0=pyhd8ed1ab_0
jupyter_server_terminals=0.5.3=pyhd8ed1ab_0
jupyterlab=4.1.5=pyhd8ed1ab_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_1
jupyterlab_server=2.25.4=pyhd8ed1ab_0
jupyterlab_widgets=3.0.10=pyhd8ed1ab_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h166bdaf_1003
lazy-loader=0.3=pypi_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libabseil=20240116.2=cxx17_h59595ed_0
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.0.20=0
libcurand=10.3.5.119=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.19=hd590300_0
libexpat=2.6.2=h59595ed_0
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran-ng=13.2.0=h69a702a_5
libgfortran5=13.2.0=ha4646dd_5
libgomp=14.2.0=h77fa898_1
libgrpc=1.62.2=h15f2491_0
libhwloc=2.9.3=default_h554bfaf_1009
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=16_linux64_mkl
liblapacke=3.9.0=16_linux64_mkl
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.43=h2797004_0
libprotobuf=4.25.3=h08a7969_0
libre2-11=2023.09.01=h5a48ba9_2
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.45.2=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_5
libtiff=4.6.0=ha9c0a0a_2
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.2=hd590300_0
libxcb=1.15=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.12.6=h232c23b_0
libzlib=1.2.13=hd590300_5
llvm-openmp=15.0.7=h0cdce71_0
markdown=3.6=pyhd8ed1ab_0
markupsafe=2.1.5=py312h98912ed_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=3.0.2=pyhd8ed1ab_0
mkl=2022.1.0=h84fe81f_915
mkl-devel=2022.1.0=ha770c72_916
mkl-include=2022.1.0=h84fe81f_915
mpmath=1.3.0=pyhd8ed1ab_0
nbclient=0.10.0=pyhd8ed1ab_0
nbconvert=7.16.2=pyhd8ed1ab_0
nbconvert-core=7.16.2=pyhd8ed1ab_0
nbconvert-pandoc=7.16.2=pyhd8ed1ab_0
nbformat=5.10.3=pyhd8ed1ab_0
ncurses=6.4.20240210=h59595ed_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
networkx=3.2.1=pyhd8ed1ab_0
nibabel=5.2.1=pyha770c72_0
notebook=7.1.2=pyhd8ed1ab_0
notebook-shim=0.2.4=pyhd8ed1ab_0
numpy=1.26.4=py312heda63a1_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.2=h488ebb8_0
openssl=3.4.0=hb9d3cd8_0
overrides=7.7.0=pyhd8ed1ab_0
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.1=py312hfb8ada1_0
pandoc=3.1.12.3=ha770c72_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=py312hf3581a9_0
pip=24.3.1=pyh8b19718_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.2.0=pyhd8ed1ab_0
plotly=5.24.1=py_0
prometheus_client=0.20.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=4.25.3=py312h72fbbdf_0
psutil=5.9.8=py312h98912ed_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.12.2=hab00c5b_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-fastjsonschema=2.19.1=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pytorch=2.2.1=py3.12_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py312h98912ed_1
pyzmq=25.1.2=py312h886d080_0
qtconsole-base=5.5.1=pyha770c72_0
qtpy=2.4.1=pyhd8ed1ab_0
re2=2023.09.01=h7f4b329_2
readline=8.2=h8228510_1
referencing=0.34.0=pyhd8ed1ab_0
requests=2.31.0=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rpds-py=0.18.0=py312h4b3b743_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
send2trash=1.8.2=pyh41d4057_0
setuptools=69.2.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pyhd8ed1ab_0
soupsieve=2.5=pyhd8ed1ab_1
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pyh04b8f61_3
tbb=2021.11.0=h00ab1b0_1
tenacity=8.3.0=pyhd8ed1ab_0
tensorboard=2.16.2=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py312h241aef2_1
terminado=0.18.1=pyh0d859eb_0
threadpoolctl=3.4.0=pypi_0
tifffile=2024.2.12=pypi_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=2.2.1=py312_cu121
torchvision=0.17.1=py312_cu121
tornado=6.4=py312h98912ed_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pyhd8ed1ab_0
types-python-dateutil=2.9.0.20240316=pyhd8ed1ab_0
typing-extensions=4.10.0=hd8ed1ab_0
typing_extensions=4.10.0=pyha770c72_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2024a=h0c530f3_0
uri-template=1.3.0=pyhd8ed1ab_0
urllib3=2.2.1=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=1.13=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.7.0=pyhd8ed1ab_0
werkzeug=3.0.3=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_0
widgetsnbextension=4.0.10=pyhd8ed1ab_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.5=h59595ed_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
